+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Set `options(bitmapType = 'cairo')` only when the system supports it. Previously only `capabilities('cairo')` was checked, which was not enough (e.g., on macOS without xquartz installed). This option can also be manually set via `options(litedown.bitmapType)` if the automatic switch to `cairo` is not desired.
 
+- Fixed the bug that indented or quoted code blocks are not correctly indented or quoted when a code expression contains multiple lines.
+
 - Fixed the bug that the span syntax `[text](){...}` doesn't work when `text` contains markup (e.g., bold or italic).
 
 # CHANGES IN litedown VERSION 0.3